requirements.txt
pyserial==3.5
pyttsx3==2.90
pydub==0.25.1
gtts
ffmpeg

